pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tok_library"
version = "0.1.0"
description = "Tokenization library"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "transformers",
    "datasets",
    "pyyaml",
    "pandas",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "flake8",
]

[tool.setuptools]
package-dir = {"" = "src"}  # Points to the src directory where your code resides